requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
crewai==0.1.0
pinecone-client==2.2.4
openai==1.3.7
beautifulsoup4==4.12.2
requests==2.31.0
scrapy==2.11.0
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
sendgrid==6.10.0
schedule==1.2.0
python-dateutil==2.8.2
aiofiles==23.2.1
httpx==0.25.2